payment-service/pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "psycopg2-binary>=2.9.11",
]
